Clean up coprocessor settings printing and fix network info
